requirements.txt
torch==2.5.1
tqdm==4.67.1
torchvision==0.20.1
matplotlib==3.10.0
kagglehub==0.3.6
numpy==1.26.4
pillow==11.1.0
scikit-image==0.25.1